move functions from AutoStory to AutoStoryTools
